Reorder functions related to article reset
